DHFPROD-7767: Write HC models to draft and add publish endpoint
