+++ installer-for-dhs/src/main/java/com/marklogic/hub/dhs/installer/command/AbstractVerifyCommand.java
@@ -110,7 +110,8 @@ public abstract class AbstractVerifyCommand extends AbstractInstallerCommand {
         for (String triggersDatabaseName : new String[]{"data-hub-final-TRIGGERS", "data-hub-staging-TRIGGERS"}) {
             ResourcesFragment triggers = new TriggerManager(hubConfig.getManageClient(), triggersDatabaseName).getAsXml();
             for (String trigger : new String[]{"ml-dh-entity-create", "ml-dh-entity-delete", "ml-dh-entity-modify",
-                "ml-dh-entity-validate-create", "ml-dh-entity-validate-modify"}) {
+                "ml-dh-entity-validate-create", "ml-dh-entity-validate-modify",
+                    "ml-dh-draft-entity-validate-create", "ml-dh-draft-entity-validate-modify"}) {
                 verify(triggers.resourceExists(trigger), "Expected trigger to be created in " + triggersDatabaseName + ": " + trigger);
             }
         }
