+++ marklogic-data-hub/src/main/resources/hub-internal-config/triggers/ml-dh-draft-entity-validate-modify.json
@@ -1,0 +1,39 @@
+{
+  "name": "ml-dh-draft-entity-validate-modify",
+  "description": "MarkLogic Data Hub draft entity model update validation pre-commit trigger",
+  "event": {
+    "data-event": {
+      "collection-scope": {
+        "uri": "http://marklogic.com/entity-services/models/draft"
+      },
+      "document-content": {
+        "update-kind": "modify"
+      },
+      "when": "pre-commit"
+    }
+  },
+  "module": "data-hub/4/triggers/entity-model-validate-trigger.xqy",
+  "module-db": "%%mlModulesDbName%%",
+  "module-root": "/",
+  "enabled": true,
+  "recursive": true,
+  "task-priority": "normal",
+  "permission": [
+    {
+      "role-name": "%%mlFlowDeveloperRole%%",
+      "capability": "update"
+    },
+    {
+      "role-name": "%%mlFlowOperatorRole%%",
+      "capability": "read"
+    },
+    {
+      "role-name": "data-hub-developer",
+      "capability": "update"
+    },
+    {
+      "role-name": "data-hub-common",
+      "capability": "read"
+    }
+  ]
+}
