+++ marklogic-data-hub/src/main/resources/ml-modules/root/data-hub/5/data-services/models/publishDraftModels.api
@@ -1,0 +1,5 @@
+{
+  "functionName": "publishDraftModels",
+  "desc": "Moves draft entity models to the published collection and clear out the draft collection",
+  "params": []
+}
