+++ marklogic-data-hub/src/main/resources/ml-modules/root/data-hub/5/data-services/models/updateDraftModelEntityTypes.api
@@ -1,5 +1,6 @@
 {
-  "functionName": "updateModelEntityTypes",
+  "functionName": "updateDraftModelEntityTypes",
+  "desc" : "Update entity model types in the entity models draft collection.",
   "params": [
     {
       "name": "input",
