+++ marklogic-data-hub/src/main/resources/ml-modules/root/data-hub/5/impl/consts.sjs
@@ -16,6 +16,7 @@
 
 const FLOW_COLLECTION = "http://marklogic.com/data-hub/flow";
 const ENTITY_MODEL_COLLECTION = "http://marklogic.com/entity-services/models";
+const DRAFT_ENTITY_MODEL_COLLECTION = `${ENTITY_MODEL_COLLECTION}/draft`;
 const STEP_DEFINITION_COLLECTION = "http://marklogic.com/data-hub/step-definition";
 const STEP_COLLECTION = "http://marklogic.com/data-hub/steps";
 
@@ -71,6 +72,7 @@ module.exports = {
 
   FLOW_COLLECTION,
   ENTITY_MODEL_COLLECTION,
+  DRAFT_ENTITY_MODEL_COLLECTION,
   STEP_DEFINITION_COLLECTION,
   STEP_COLLECTION,
